Moved LoadRandomExternalKey functionality into the internal package.
